camelCase is bit confusing it can be kebabCase
Document Change Request : camelCase is bit confusing it can be kebabCase

While reading the doc I felt like the doc shall be edited as the component is made using 'camelCase' where as in html it is referenced as 'kebab-case'. It shall be 'kebabCase' as component name and in html 'kebab-case'

I feel it shall be changed as its confusing.